+++ pkg/purgecache/errors.go
@@ -14,6 +14,9 @@ import (
 var (
 	// ErrStructValidation is returned when given struct validation failed.
 	ErrStructValidation = errors.New("struct validation")
+
+	// ErrRateLimitStatus is returned when RateLimitStatus fails.
+	ErrRateLimitStatus = errors.New("checking rate limit status")
 )
 
 type (
@@ -24,18 +27,18 @@ type (
 		Detail                      string `json:"detail"`
 		DescribedBy                 string `json:"describedBy"`
 		SupportID                   string `json:"supportId"`
-		RateLimit                   int    `json:"rateLimit,omitempty"`
-		RateLimitCurrentRequestSize int    `json:"rateLimitCurrentRequestSize,omitempty"`
-		RateLimitRemaining          int    `json:"rateLimitRemaining,omitempty"`
+		RateLimit                   *int64 `json:"rateLimit,omitempty"`
+		RateLimitCurrentRequestSize *int64 `json:"rateLimitCurrentRequestSize,omitempty"`
+		RateLimitRemaining          *int64 `json:"rateLimitRemaining,omitempty"`
 	}
 )
 
 // Error parses an error from the Purge Cache API response.
-func (r *purgecache) Error(resp *http.Response) error {
+func (p *purgecache) Error(resp *http.Response) error {
 	var e Error
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		r.Log(resp.Request.Context()).Errorf("reading error response body: %s", err)
+		p.Log(resp.Request.Context()).Errorf("reading error response body: %s", err)
 		e.HTTPStatus = resp.StatusCode
 		e.Title = "Failed to read error body"
 		e.Detail = err.Error()
@@ -43,7 +46,7 @@ func (r *purgecache) Error(resp *http.Response) error {
 	}
 
 	if err := json.Unmarshal(body, &e); err != nil {
-		r.Log(resp.Request.Context()).Errorf("could not unmarshal API error: %s", err)
+		p.Log(resp.Request.Context()).Errorf("could not unmarshal API error: %s", err)
 		e.HTTPStatus = resp.StatusCode
 		e.Title = "Failed to unmarshal error body. Purge Cache API failed. Check details for more information."
 		e.Detail = errs.UnescapeContent(string(body))